separate policy gradient example with augmentation in another file
